requirements.txt
aiohttp==3.8.4
nextcord==2.4.2
marshmallow==3.19.0
packaging==23.1
motor==3.1.2
pymongo==4.3.3
dnspython==2.3.0
typing-extensions==4.6.2
onami==2.6.3
braceexpand==0.1.7
click==8.1.3
import_expression==1.1.4
line_profiler==4.0.3
colorama==0.4.6
astunparse==1.6.3
wheel==0.40.0
six==1.16.0
python-dotenv==1.0.0